http://wesleycai.com/math/EMS_Paper.pdf
Abstract:
We present a method to optimize a shift schedule for each station as
well as finding the most optimal
way to place paramedics and ambulances at stations located in and
around the Boston metropolitan area.
The goal is to minimize the need for ambulances during a specific time
period, given a certain expected
coverage. Using data received from the Boston Emergency Services
regarding the number of incidents per
hour in a day, we generated a demand curve for our study area Roxbury.
We plan to use the model of the
study area to analyze other parts of Boston. The end goal is to compare
our model with the linear model in
Ragopalan et. al paper Ambulance Deployment and Shift Scheduling: An
Integrated Approach. Our final
results were produced from a simulation that was ran 10,000 times which
represent the simulated days. It
gave us a random the number of calls per day and a random number of
ambulances that would be needed.
We ran the code in R, which is a statistical program.